refactor: userRatingCount의 자료형을 Double로 변경
- 1,000단위로 문자열 변환 처리를 위함

+++ AppStore/AppStore/Data/Network/ResponseDTO.swift
@@ -19,7 +19,7 @@ struct AppResponse: Decodable {
     let appCategory: String
     let price: Double
     let formattedPrice: String
-    let userRatingCount: Int
+    let userRatingCount: Double
     let averageUserRating: Double
     let screenshotURLs: [String]
     let contentAdvisoryRating: String
