+++ AppStore/AppStore/Domain/Entity/App.swift
@@ -12,7 +12,7 @@ struct App {
     let appCategory: String
     let price: Double
     let formattedPrice: String
-    let userRatingCount: Int
+    let userRatingCount: Double
     let averageUserRating: Double
     let screenshotURLs: [String]
     let contentAdvisoryRating: String
